[#861] file renaming
